control feature visibility with two uniforms: currentTimeMin/Max and an attribute time
When a feature changes OL has to rebuild the buffers and upload them to
the GPU. This takes time, so it would be nice if we minimize this.
We already passed all features to OL and thus WebGL. But instead of
controlling visibility by the attribute opacity we use the constant time
attribute on the feature and use two uniforms to control the minimum
and maximum time we want to show. Changing the time and thus the
visibility of features is now instant (60 fps when sliding maxTime like
a madman).